Consolidate fakers and their models
