+++ src/Models/JobModel.php
@@ -1,6 +1,8 @@
 <?php namespace Tatter\Workflows\Models;
 
 use CodeIgniter\Model;
+use CodeIgniter\Test\Fabricator;
+use Faker\Generator;
 use Tatter\Workflows\Entities\Job;
 
 class JobModel extends Model
@@ -89,5 +91,22 @@ class JobModel extends Model
 		}
 
 		return $data;
+	}
+
+	/**
+	 * Faked data for Fabricator.
+	 *
+	 * @param Generator $faker
+	 *
+	 * @return Job
+	 */
+	public function fake(Generator &$faker): Job
+	{
+		return new Job([
+			'name'        => $faker->catchPhrase,
+			'summary'     => $faker->sentence,
+			'workflow_id' => rand(1, Fabricator::getCount('workflows') ?: 4),
+			'stage_id'    => rand(1, Fabricator::getCount('stages') ?: 99),
+		]);
 	}
 }
