+++ src/Models/StageModel.php
@@ -1,6 +1,8 @@
 <?php namespace Tatter\Workflows\Models;
 
 use CodeIgniter\Model;
+use CodeIgniter\Test\Fabricator;
+use Faker\Generator;
 use Tatter\Workflows\Entities\Action;
 use Tatter\Workflows\Entities\Stage;
 use Tatter\Workflows\Entities\Workflow;
@@ -25,4 +27,20 @@ class StageModel extends Model
 	protected $afterInsert = ['auditInsert'];
 	protected $afterUpdate = ['auditUpdate'];
 	protected $afterDelete = ['auditDelete'];
+
+	/**
+	 * Faked data for Fabricator.
+	 *
+	 * @param Generator $faker
+	 *
+	 * @return Stage
+	 */
+	public function fake(Generator &$faker): Stage
+	{
+		return new Stage([
+			'action_id'   => rand(1, Fabricator::getCount('actions') ?: 12),
+			'workflow_id' => rand(1, Fabricator::getCount('workflows') ?: 4),
+			'required'    => (bool) rand(0, 5),
+		]);
+	}
 }
