+++ src/Models/WorkflowModel.php
@@ -1,6 +1,7 @@
 <?php namespace Tatter\Workflows\Models;
 
 use CodeIgniter\Model;
+use Faker\Generator;
 use Tatter\Workflows\Entities\Workflow;
 use Tatter\Workflows\Models\StageModel;
 
@@ -51,5 +52,23 @@ class WorkflowModel extends Model
 		}
 
 		return $result;
+	}
+
+	/**
+	 * Faked data for Fabricator.
+	 *
+	 * @param Generator $faker
+	 *
+	 * @return Workflow
+	 */
+	public function fake(Generator &$faker): Workflow
+	{
+		return new Workflow([
+			'name'        => $faker->word,
+			'category'    => $faker->streetSuffix,
+			'icon'        => $faker->safeColorName,
+			'summary'     => $faker->sentence,
+			'description' => $faker->paragraph,
+		]);
 	}
 }
