+++ tests/entities/JobflagTest.php
@@ -2,7 +2,7 @@
 
 use CodeIgniter\Database\BaseBuilder;
 use CodeIgniter\I18n\Time;
-use Tatter\Workflows\Test\Fakers\JobFaker;
+use Tatter\Workflows\Models\JobModel;
 use Tests\Support\DatabaseTestCase;
 
 class JobflagTest extends DatabaseTestCase
@@ -33,7 +33,7 @@ class JobflagTest extends DatabaseTestCase
 		parent::setUp();
 
 		$this->now     = date('Y-m-d H:i:s');
-		$this->job     = fake(JobFaker::class);
+		$this->job     = fake(JobModel::class);
 		$this->builder = $this->db->table('jobflags');
 	}
